// server/src/models/LeagueSettings.ts
export interface LeagueSettings {
    teamCount: number;
    rosterSize: number;
    qbSlots: number;
    rbSlots: number;
    wrSlots: number;
    teSlots: number;
    flexSlots: number;
    kSlots: number;
    defSlots: number;
    benchSlots: number;
    passingTouchdownPoints: number;
    rushingTouchdownPoints: number;
    receptionPoints: number;
  }